refactor: 모달 포커스 관리 로직을 재사용 가능한 hook으로 분리
- useModalFocus: 모달 포커스 저장/복원 및 초기 포커스 설정
  - 모달 열릴 때 현재 포커스 저장
  - 닫기 버튼에 자동 포커스 (지연 시간 설정 가능)
  - 모달 닫힐 때 원래 포커스 복원
- useFocusTrap: Tab 키로 모달 내부만 순환하는 포커스 트랩
  - Shift+Tab 지원
  - 모달 외부 포커스 이동 방지
- ProjectDetailModal에서 중복 코드 제거 (약 50줄 → 2줄)
