+++ src/utils/hooks.ts
@@ -9,7 +9,8 @@ import {
 } from 'firebase/auth';
 import { collection, getDocs } from 'firebase/firestore';
 import { useRouter } from 'next/navigation';
-import { useEffect, useState } from 'react';
+import type { RefObject } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { auth, db } from '../../firebaseConfig';
 
 // Firebase 에러 코드를 한국어 메시지로 변환
@@ -219,4 +220,87 @@ export const useModalClose = (
   };
 
   return { handleBackdropClick };
+};
+
+/**
+ * 모달 포커스 관리 hook (포커스 저장/복원 및 초기 포커스 설정)
+ * @param closeButtonRef - 닫기 버튼의 ref
+ * @param isOpen - 모달이 열려있는지 여부
+ * @param options - 옵션 설정
+ * @param options.focusDelay - 포커스 지연 시간 (ms, 기본값: 100)
+ * @returns previousFocusRef - 이전 포커스 요소를 저장하는 ref
+ */
+export const useModalFocus = (
+  closeButtonRef: RefObject<HTMLButtonElement | null>,
+  isOpen: boolean,
+  options?: { focusDelay?: number }
+) => {
+  const { focusDelay = 100 } = options || {};
+  const previousFocusRef = useRef<HTMLElement | null>(null);
+
+  useEffect(() => {
+    if (!isOpen) return;
+
+    // 모달이 열릴 때 현재 포커스 저장
+    previousFocusRef.current = document.activeElement as HTMLElement;
+
+    // 애니메이션 후 닫기 버튼에 포커스
+    const timer = setTimeout(() => {
+      closeButtonRef.current?.focus();
+    }, focusDelay);
+
+    return () => {
+      clearTimeout(timer);
+      // 모달이 닫힐 때 원래 포커스로 복원
+      previousFocusRef.current?.focus();
+    };
+  }, [isOpen, closeButtonRef, focusDelay]);
+
+  return { previousFocusRef };
+};
+
+/**
+ * 모달 포커스 트랩 hook (Tab 키로 모달 내부만 순환)
+ * @param modalRef - 모달 컨테이너의 ref
+ * @param isOpen - 모달이 열려있는지 여부
+ */
+export const useFocusTrap = (
+  modalRef: RefObject<HTMLElement | null>,
+  isOpen: boolean
+) => {
+  useEffect(() => {
+    if (!isOpen) return;
+
+    const handleTabKey = (event: KeyboardEvent) => {
+      if (event.key !== 'Tab') return;
+
+      const modal = modalRef.current;
+      if (!modal) return;
+
+      // 포커스 가능한 요소들 선택
+      const focusableElements = modal.querySelectorAll<HTMLElement>(
+        'button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])'
+      );
+
+      const firstElement = focusableElements[0];
+      const lastElement = focusableElements[focusableElements.length - 1];
+
+      if (event.shiftKey) {
+        // Shift + Tab
+        if (document.activeElement === firstElement) {
+          event.preventDefault();
+          lastElement?.focus();
+        }
+      } else {
+        // Tab
+        if (document.activeElement === lastElement) {
+          event.preventDefault();
+          firstElement?.focus();
+        }
+      }
+    };
+
+    document.addEventListener('keydown', handleTabKey);
+    return () => document.removeEventListener('keydown', handleTabKey);
+  }, [isOpen, modalRef]);
 };
